structure change and add docker deployment
