+++ README.md
@@ -4,47 +4,21 @@
 
 # Bootcamp CorDapp [<img src="https://raw.githubusercontent.com/corda/samples-java/master/webIDE.png" height=25 />](https://ide.corda.net/?folder=/home/coder/bootcamp-cordapp)
 
-This project is the template we will use as a basis for developing a complete CorDapp 
+This project is the template we will use as a basis for developing a complete CorDapp
 during today's bootcamp. Our CorDapp will allow the issuance of tokens onto the ledger.
 
-We'll develop the CorDapp using a test-driven approach. At each stage, you'll know your 
+We'll develop the CorDapp using a test-driven approach. At each stage, you'll know your
 CorDapp is working once it passes both sets of tests defined in `src/test/java/bootcamp`.
 
 ## Set up
 
 1. Download and install a JDK 8 JVM (minimum supported version 8u131)
-2. Download and install IntelliJ Community Edition (supported versions 2017.x and 2018.x)
+2. Download and install IntelliJ Community Edition (supported versions: prior than 2021)
 3. Download the bootcamp-cordapp repository:
 
        git clone https://github.com/corda/bootcamp-cordapp
-       
-4. Open IntelliJ. From the splash screen, click `Import Project`, select the `bootcamp—
-cordapp` folder and click `Open`
-5. Select `Import project from external model > Gradle > Next > Finish`
-6. Click `File > Project Structure…` and select the Project SDK (Oracle JDK 8, 8u131+)
 
-    i. Add a new SDK if required by clicking `New…` and selecting the JDK’s folder
-
-7. Open the `Project` view by clicking `View > Tool Windows > Project`
-8. Run the test in `src/test/java/java_bootcamp/ProjectImportedOKTest.java`. It should pass!
-
-## Links to useful resources
-
-This project contains example state, contract and flow implementations:
-
-* `src/main/java/java_examples/ArtState`
-* `src/main/java/java_examples/ArtContract`
-* `src/main/java/java_examples/ArtTransferFlowInitiator`
-* `src/main/java/java_examples/ArtTransferFlowResponder`
-
-There are also several web resources that you will likely find useful for this
-bootcamp:
-
-* Key Concepts docs (`docs.corda.net/key-concepts.html`)
-* API docs (`docs.corda.net/api-index.html`)
-* Cheat sheet (`docs.corda.net/cheat-sheet.html`)
-* Sample CorDapps (`www.corda.net/samples`)
-* Stack Overflow (`www.stackoverflow.com/questions/tagged/corda`)
+4. Open IntelliJ load the project
 
 ## What we'll be building
 
@@ -150,20 +124,28 @@ Once you've finished the CorDapp's code, run it with the following steps:
 * Build a test network of nodes by opening a terminal window at the root of
   your project and running the following command:
 
-    * Windows:   `gradlew.bat deployNodes`
-    * macOS:     `./gradlew deployNodes`
+  * Windows:   `gradlew.bat deployNodes`
+  * Linux/Mac:     `./gradlew deployNodes`
 
 * Start the nodes by running the following command:
 
-    * Windows:   `build\nodes\runnodes.bat`
-    * macOS:     `build/nodes/runnodes`
+  * Windows:   `build\nodes\runnodes.bat`
+  * Linux/Mac: `./build/nodes/runnodes`
+
+
+
+* Build and Start the nodes using docker:
+  * Windows:   `gradlew.bat prepareDockerNodes`
+  * Linux/Mac:     `./gradlew prepareDockerNodes`
+  * `ACCEPT_LICENSE=Y | docker-compose -f ./build/nodes/docker-compose.yml up`
+
 
 * Open the nodes are started, go to the terminal of Party A (not the notary!)
   and run the following command to issue 99 tokens to Party B:
 
-    `flow start TokenIssueFlow owner: PartyB, amount: 99`
+  `flow start TokenIssueFlowInitiator owner: PartyB, amount: 99`
 
-* You can now see the tokens in the vaults of Party A and Party B (but not 
+* You can now see the tokens in the vaults of Party A and Party B (but not
   Party C!) by running the following command in their respective terminals:
 
-    `run vaultQuery contractStateType: bootcamp.TokenState`
+  `run vaultQuery contractStateType: com.bootcamp.contracts.TokenState`
