+++ gradle.properties
@@ -1,3 +1,3 @@
-name=Bootcamp
+name=Test
 group=com.bootcamp
 version=0.1
