+++ repositories.gradle
@@ -1,7 +1,6 @@
 repositories {
     mavenLocal()
     mavenCentral()
-    jcenter()
     maven { url 'https://jitpack.io' }
     maven { url 'https://software.r3.com/artifactory/corda' }
     maven { url 'https://repo.gradle.org/gradle/libs-releases' }
