+++ settings.gradle
@@ -1,0 +1,2 @@
+include 'workflows'
+include 'contracts'
