--- src/main/java/bootcamp/TokenIssueFlowInitiator.java
@@ -1,69 +1,0 @@
-package bootcamp;
-
-import co.paralleluniverse.fibers.Suspendable;
-import net.corda.core.flows.*;
-import net.corda.core.identity.Party;
-import net.corda.core.transactions.SignedTransaction;
-import net.corda.core.transactions.TransactionBuilder;
-import net.corda.core.utilities.ProgressTracker;
-import net.corda.core.contracts.CommandData;
-
-import static java.util.Collections.singletonList;
-
-@InitiatingFlow
-@StartableByRPC
-public class TokenIssueFlowInitiator extends FlowLogic<SignedTransaction> {
-    private final Party owner;
-    private final int amount;
-
-    public TokenIssueFlowInitiator(Party owner, int amount) {
-        this.owner = owner;
-        this.amount = amount;
-    }
-
-    private final ProgressTracker progressTracker = new ProgressTracker();
-
-    @Override
-    public ProgressTracker getProgressTracker() {
-        return progressTracker;
-    }
-
-    @Suspendable
-    @Override
-    public SignedTransaction call() throws FlowException {
-        // We choose our transaction's notary (the notary prevents double-spends).
-        Party notary = getServiceHub().getNetworkMapCache().getNotaryIdentities().get(0);
-        // We get a reference to our own identity.
-        Party issuer = getOurIdentity();
-
-        /* ============================================================================
-         *         TODO 1 - Create our TokenState to represent on-ledger tokens!
-         * ===========================================================================*/
-        // We create our new TokenState.
-        TokenState tokenState = null;
-
-
-        /* ============================================================================
-         *      TODO 3 - Build our token issuance transaction to update the ledger!
-         * ===========================================================================*/
-        // We build our transaction.
-        TransactionBuilder transactionBuilder = null;
-
-        /* ============================================================================
-         *          TODO 2 - Write our TokenContract to control token issuance!
-         * ===========================================================================*/
-        // We check our transaction is valid based on its contracts.
-        transactionBuilder.verify(getServiceHub());
-
-        FlowSession session = initiateFlow(owner);
-
-        // We sign the transaction with our private key, making it immutable.
-        SignedTransaction signedTransaction = getServiceHub().signInitialTransaction(transactionBuilder);
-
-        // The counterparty signs the transaction
-        SignedTransaction fullySignedTransaction = subFlow(new CollectSignaturesFlow(signedTransaction, singletonList(session)));
-
-        // We get the transaction notarised and recorded automatically by the platform.
-        return subFlow(new FinalityFlow(fullySignedTransaction, singletonList(session)));
-    }
-}
