--- src/main/java/bootcamp/TokenIssueFlowResponder.java
@@ -1,29 +1,0 @@
-package bootcamp;
-
-import co.paralleluniverse.fibers.Suspendable;
-import net.corda.core.flows.*;
-import net.corda.core.transactions.SignedTransaction;
-
-@InitiatedBy(TokenIssueFlowInitiator.class)
-public class TokenIssueFlowResponder extends FlowLogic<Void> {
-
-    private final FlowSession otherSide;
-
-    public TokenIssueFlowResponder(FlowSession otherSide) {
-        this.otherSide = otherSide;
-    }
-
-    @Override
-    @Suspendable
-    public Void call() throws FlowException {
-        SignedTransaction signedTransaction = subFlow(new SignTransactionFlow(otherSide) {
-            @Suspendable
-            @Override
-            protected void checkTransaction(SignedTransaction stx) throws FlowException {
-                // Implement responder flow transaction checks here
-            }
-        });
-        subFlow(new ReceiveFinalityFlow(otherSide, signedTransaction.getId()));
-        return null;
-    }
-}
