--- src/main/java/bootcamp/TokenState.java
@@ -1,17 +1,0 @@
-package bootcamp;
-
-import net.corda.core.contracts.BelongsToContract;
-import net.corda.core.contracts.ContractState;
-import net.corda.core.identity.AbstractParty;
-import net.corda.core.identity.Party;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/* Our state, defining a shared fact on the ledger.
- * See src/main/java/examples/ArtState.java for an example. */
-//@BelongsToContract(TokenContract.class)
-public class TokenState {
-
-}
