--- src/test/java/ProjectImportedOKTest.java
@@ -1,8 +1,0 @@
-import org.junit.Test;
-
-public class ProjectImportedOKTest {
-    @Test
-    public void projectImportedOKTest() {
-
-    }
-}
